clean refactor
