--- runway_for_ml.egg-info/PKG-INFO
@@ -1,12 +1,0 @@
-Metadata-Version: 2.1
-Name: runway-for-ml
-Version: 0.0.1
-Summary: A lightweight ML research framework
-Home-page: UNKNOWN
-Author: Jinghong Chen & Weizhe Lin
-Author-email: jc2124@cam.ac.uk
-License: UNKNOWN
-Platform: UNKNOWN
-
-UNKNOWN
-
