--- runway_for_ml.egg-info/SOURCES.txt
@@ -1,13 +1,0 @@
-README.md
-setup.py
-runway_for_ml.egg-info/PKG-INFO
-runway_for_ml.egg-info/SOURCES.txt
-runway_for_ml.egg-info/dependency_links.txt
-runway_for_ml.egg-info/top_level.txt
-tests/__init__.py
-utils/__init__.py
-utils/cache_system.py
-utils/config_system.py
-utils/dirs.py
-utils/imports.py
-utils/mixin_utils.py
