--- runway_for_ml.egg-info/top_level.txt
@@ -1,2 +1,0 @@
-tests
-utils
