+++ utils/global_variables.py
